Revert "chore(blog): move post mdx parser to lib (#289)"
